requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-classic
langchain-groq
duckduckgo-search
ddgs
arxiv
wikipedia
python-dotenv
